requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.1
python-dotenv>=1.0.1
redis>=5.0.1
celery>=5.3.4
requests>=2.31.0
